requirements.txt
python-telegram-bot==20.8
qbittorrent-api==2023.10.52
apscheduler==3.10.4
